% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_lib.R
\name{AccumulateInteractionEffects}
\alias{AccumulateInteractionEffects}
\title{Accumulate responses of group membership interactions}
\usage{
AccumulateInteractionEffects(group_effects)
}
\arguments{
\item{group_effects}{A list of group effects, e.g. as returned by \code{DrawGroupEffects}}
}
\value{
A dataframe with regressors, groups, and ey values suitable for
joining with raw data.
}
\description{
Accumulate responses of group membership interactions
}
