% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_lib.R
\name{AggregateSimulationData}
\alias{AggregateSimulationData}
\title{Aggregate simulation data by group.}
\usage{
AggregateSimulationData(sim_data)
}
\arguments{
\item{sim_data}{Output of \code{SimulateSurveyData}}
}
\value{
A list containing aggregated and joined simulation data
}
\description{
Aggregate simulation data by group.
}
