% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_lib.R
\name{CheckCovariateBalance}
\alias{CheckCovariateBalance}
\title{Use MrPaw to check covariate balance.}
\usage{
CheckCovariateBalance(mrpaw_list, x_sur, x_pop, pop_w = NULL)
}
\arguments{
\item{mrpaw_list}{The output of one of the Get*MCMCWeights functions}

\item{x_sur}{A matrix of some regressors in the survey data}

\item{x_pop}{A matrix of the same regressors in the population data}

\item{pop_w}{(Optional) A vector of population weights. Taken to be all one if NULL.}
}
\value{
The weighted average regressors in each regressor matrix, where the
weights in the survey are given by the MrP affine weights.
The regressor matrices might be made with a call to model.matrix using the
same formula for each dataset.
}
\description{
Use MrPaw to check covariate balance.
}
