% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_lib.R
\name{CheckDatapointWeighting}
\alias{CheckDatapointWeighting}
\title{Use MrPaw to check invariance to weighting datapoints by covariate
(i.e. distribution shift aligned with a covariate).}
\usage{
CheckDatapointWeighting(infl_list, x_sur)
}
\arguments{
\item{infl_list}{The output of the EvalInfluenceFunction function}

\item{x_sur}{A matrix of some regressors in the survey data}
}
\value{
The derivative of MrP in the direction of a distribution
shift aligned with each regressor in x_sur.
}
\description{
Use MrPaw to check invariance to weighting datapoints by covariate
(i.e. distribution shift aligned with a covariate).
}
