% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_lib.R
\name{EvalAlphaHat}
\alias{EvalAlphaHat}
\title{Estimate the coefficient were extra regressors to be included.  (Experimental)}
\usage{
EvalAlphaHat(x0, mrpaw_list, post, survey_df)
}
\arguments{
\item{x0}{A matrix whose coefficients are assumed to be zero in the original fit}

\item{mrpaw_list}{The output of one of the Get*MCMCWeights functions}

\item{survey_df}{The survey dataframe}
}
\value{
The estimated change (implicitly from zero) in the estimated
coefficients using a single Newton step at each draw.
}
\description{
Estimate the coefficient were extra regressors to be included.  (Experimental)
}
