% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_lib.R
\name{EvalFisherConsistency}
\alias{EvalFisherConsistency}
\title{Evaluate the Fisher consistency of the coefficients.
Assume the data are drawn from the correct parametric model, and the
posterior means are consistent.  Then if we use importance sampling to
weight the log likelihood with parameter theta_1 versus theta_0, then
the change in the estimate of theta should be (theta_1 - theta_0), and
the derivative with respect to this change should be 1.}
\usage{
EvalFisherConsistency(mrpaw_list, post, survey_df)
}
\arguments{
\item{mrpaw_list}{The output of one of the Get*MCMCWeights functions}

\item{post}{The output of \code{brm(..., survey_df, family=binomial(link="logit"))}}

\item{survey_df}{The survey dataframe}
}
\description{
Evaluate the Fisher consistency of the coefficients.
Assume the data are drawn from the correct parametric model, and the
posterior means are consistent.  Then if we use importance sampling to
weight the log likelihood with parameter theta_1 versus theta_0, then
the change in the estimate of theta should be (theta_1 - theta_0), and
the derivative with respect to this change should be 1.
}
