% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_lib.R
\name{EvalInfluenceFunction}
\alias{EvalInfluenceFunction}
\title{Evaluate the nonparametric influence function and IJ variance.}
\usage{
EvalInfluenceFunction(mrpaw_list, post, survey_df)
}
\arguments{
\item{mrpaw_list}{The output of one of the Get*MCMCWeights functions}

\item{post}{The output of \code{brm(..., survey_df, family=binomial(link="logit"))}}

\item{survey_df}{The survey dataframe}
}
\description{
Evaluate the nonparametric influence function and IJ variance.
}
