% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_lib.R
\name{EvalOffset}
\alias{EvalOffset}
\title{Evaluate the constant in the local expansion.}
\usage{
EvalOffset(mrpaw_list, y)
}
\arguments{
\item{mrpaw_list}{The output of one of the Get*Weights functions}

\item{y}{The original vector of responses}
}
\description{
Evaluate the constant in the local expansion.
}
