% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_lib.R
\name{EvalParametricBootstrap}
\alias{EvalParametricBootstrap}
\title{Evaluate the linear approximation to the parametric bootstrap bias and variance.}
\usage{
EvalParametricBootstrap(mrpaw_list, post, survey_df)
}
\arguments{
\item{mrpaw_list}{The output of one of the Get*MCMCWeights functions}

\item{post}{The output of \code{brm(..., survey_df, family=binomial(link="logit"))}}

\item{survey_df}{The survey dataframe}
}
\description{
Evaluate the linear approximation to the parametric bootstrap bias and variance.
}
