% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lib.R
\name{GetBlockBootstrapCovarianceDraws}
\alias{GetBlockBootstrapCovarianceDraws}
\title{Estimate Monte Carlo standard errors of sample covariances or
by block bootstrapping draws from an MCMC chain.}
\usage{
GetBlockBootstrapCovarianceDraws(
  draws1_mat,
  draws2_mat,
  num_blocks,
  num_draws,
  show_progress_bar = FALSE
)
}
\arguments{
\item{draws1_mat}{One set of parameter draws.}

\item{draws2_mat}{Another set of parameter draws.}

\item{num_blocks}{The number of blocks in the block bootstrap.}

\item{num_draws}{The number of bootstrap draws.}

\item{show_progress_par.}{Optional.  If TRUE, show a progress bar.
By default, FALSE.}
}
\value{
A list containing the draws of the covariance cov_samples
and the estimated Monte Carlo sample errors in cov_se.
}
\description{
Estimate Monte Carlo standard errors of sample covariances or
by block bootstrapping draws from an MCMC chain.
}
