% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_lib.R
\name{GetLogitImportanceWeights}
\alias{GetLogitImportanceWeights}
\title{Evaluate importance sampling MCMC weights for new y values.}
\usage{
GetLogitImportanceWeights(logit_post, survey_df, y_new)
}
\arguments{
\item{logit_post}{The output of \code{brm(..., survey_df, family=binomial(link="logit"))}}

\item{survey_df}{The survey dataframe}

\item{y_new}{A vector of new responses, the same length as the original response}
}
\description{
Evaluate importance sampling MCMC weights for new y values.
}
