% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_lib.R
\name{GetOLSImportanceWeights}
\alias{GetOLSImportanceWeights}
\title{Evaluate importance sampling MCMC weights for new y values.}
\usage{
GetOLSImportanceWeights(lin_post, survey_df, y_new)
}
\arguments{
\item{lin_post}{The output of \code{brm(..., survey_df, family=gaussian())}}

\item{survey_df}{The survey dataframe}

\item{y_new}{A vector of new responses, the same length as the original response}
}
\description{
Evaluate importance sampling MCMC weights for new y values.
}
