% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights_opt_lib.R
\name{GetOLSWeights}
\alias{GetOLSWeights}
\title{Get MrPaw weights for the OLS estimator.}
\usage{
GetOLSWeights(lm_fit, survey_df, pop_df, pop_w = NULL)
}
\arguments{
\item{lm_fit}{The output of \code{lm}}

\item{survey_df}{The survey dataframe}

\item{pop_df}{The population dataframe}
}
\value{
The MrP OLS estimate for sanity checking, and the weight vector
whose n-th entry is d MrP / d y_n.
}
\description{
Get MrPaw weights for the OLS estimator.
}
