% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_lib.R
\name{DrawResponse}
\alias{DrawResponse}
\title{Draw binary responses for a particular set of group memberships}
\usage{
DrawResponse(g_matrix, group_effects)
}
\arguments{
\item{g_matrix}{A matrix of group memberships, e.g. as returned by \code{DrawCorrelatedGroups}}

\item{group_effects}{A list of group effects, e.g. as returned by \code{DrawGroupEffects}}
}
\value{
A dataframe containing group memberships \code{s} and responses \code{y} corresponding
to the rows of \code{g_matrix}.
}
\description{
Draw binary responses for a particular set of group memberships
}
